% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen.CI.R
\name{cohen.CI}
\alias{cohen.CI}
\title{Function to compute CI around Cohen's d effect size}
\usage{
cohen.CI(
  m1,
  m2,
  sd1,
  sd2,
  n1,
  n2,
  conf.level,
  var.equal,
  unbiased,
  alternative,
  na.rm
)
}
\arguments{
\item{m1}{the average score of the first group}

\item{m2}{the average score of the second group}

\item{sd1}{the standard deviation the first group}

\item{sd2}{the standard deviation the second group}

\item{n1}{the first sample size}

\item{n2}{the second sample size}

\item{conf.level}{confidence level of the interval}

\item{var.equal}{a logical variable indicating whether to assume equality of population variances.
If TRUE the pooled variance is used to estimate the standard error (= Cohen's d or Hedges' g). Otherwise, the square root of the non pooled
average of both variance estimates is used to estimate the standard error (Cohen's d' or Hedges' g').}

\item{unbiased}{a logical variable indicating whether to compute the biased or unbiased estimator.
If TRUE, unbiased estimator is computed (Hedges' g or Hedges' g'). Otherwise, bias estimator is computed (Cohen's d or Cohen's d').}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{na.rm}{set whether Missing Values should be excluded (na.rm = TRUE) or not (na.rm = FALSE) - defaults to TRUE.}
}
\value{
Returns Cohen's estimators of effect size and (1-alpha)\% confidence interval around it, standard error
}
\description{
Function to compute CI around Cohen's d effect size
}
\keyword{Cohen's}
\keyword{confidence}
\keyword{effect}
\keyword{interval}
\keyword{sizes,}
