% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacohen_CI.R
\name{datacohen_CI}
\alias{datacohen_CI}
\title{Function to compute CI around Cohen's effect size estimators}
\usage{
datacohen_CI(
  Group.1,
  Group.2,
  conf.level,
  var.equal,
  unbiased,
  alternative,
  na.rm
)
}
\arguments{
\item{Group.1}{a (non-empty) numeric vector of data values.}

\item{Group.2}{a (non-empty) numeric vector of data values.}

\item{conf.level}{confidence level of the interval}

\item{var.equal}{a logical variable indicating whether to assume equality of population variances.
If TRUE the pooled variance is used to estimate the standard error (= Cohen's d or Hedges' g). Otherwise, the square root of the non pooled
average of both variance estimates is used to estimate the standard error (Cohen's d' or Hedges' g').}

\item{unbiased}{a logical variable indicating whether to compute the biased or unbiased estimator.
If TRUE, unbiased estimator is computed (Hedges' g or Hedges' g'). Otherwise, bias estimator is computed (Cohen's d or Cohen's d').}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{na.rm}{set whether Missing Values should be excluded (na.rm = TRUE) or not (na.rm = FALSE) - defaults to TRUE.}
}
\value{
Returns Cohen's estimators of effect size and (1-alpha)\% confidence interval around it, standard error
}
\description{
Function to compute CI around Cohen's effect size estimators
}
\keyword{Cohen's}
\keyword{confidence}
\keyword{effect}
\keyword{interval}
\keyword{sizes,}
