% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataglass.CI.R
\name{dataglass.CI}
\alias{dataglass.CI}
\title{Function to compute CI around Glass's effect size estimator}
\usage{
dataglass.CI(Group.1, Group.2, conf.level, unbiased, alternative, na.rm)
}
\arguments{
\item{Group.1}{a (non-empty) numeric vector of data values for control group.}

\item{Group.2}{a (non-empty) numeric vector of data values for experimental group.}

\item{conf.level}{confidence level of the interval}

\item{unbiased}{a logical variable indicating whether to compute the biased or unbiased estimator.
If TRUE, unbiased estimator is computed (Glass' g). Otherwise, bias estimator is computed (Glass's d).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{na.rm}{set whether Missing Values should be excluded (na.rm = TRUE) or not (na.rm = FALSE) - defaults to TRUE.}
}
\value{
Returns Glass's estimator of effect size and (1-alpha)\% confidence interval around it, standard error
}
\description{
Function to compute CI around Glass's effect size estimator
}
\keyword{Glass's}
\keyword{confidence}
\keyword{effect}
\keyword{interval}
\keyword{sizes,}
