% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{glass}
\alias{glass}
\title{Function to compute CI around Glass's effect size estimators}
\usage{
glass(m1, m2, sd1, sd2, n1, n2, conf.level, unbiased, alternative)
}
\arguments{
\item{m1}{the average score of the control group}

\item{m2}{the average score of the experimental group}

\item{sd1}{the standard deviation the control group}

\item{sd2}{the standard deviation the experimental group}

\item{n1}{the sample size of the control group}

\item{n2}{the sample size of the experimental group}

\item{conf.level}{confidence level of the interval}

\item{unbiased}{a logical variable indicating whether to compute the biased or unbiased estimator.
If TRUE, unbiased estimator is computed (Glass's g). Otherwise, bias estimator is computed (Glass's d).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
Returns glass's estimators of effect size and (1-alpha)\% confidence interval around it, standard error
}
\description{
Function to compute CI around Glass's effect size estimators
}
\keyword{confidence}
\keyword{effect}
\keyword{glass's}
\keyword{interval}
\keyword{sizes,}
