% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandiff_CI.R
\name{meandiff_CI}
\alias{meandiff_CI}
\title{Function to compute CI around the raw mean difference}
\usage{
meandiff_CI(m1, m2, sd1, sd2, n1, n2, conf.level, var.equal, alternative)
}
\arguments{
\item{m1}{the average score of the first group}

\item{m2}{the average score of the second group}

\item{sd1}{the standard deviation the first group}

\item{sd2}{the standard deviation the second group}

\item{n1}{the first sample size}

\item{n2}{the second sample size}

\item{conf.level}{confidence level of the interval}

\item{var.equal}{a logical variable indicating whether to assume equality of population variances.
If TRUE the pooled variance is used to estimate the standard error. Otherwise, the standard error is estimated based on
unpooled variance.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
Returns raw mean difference, (1-alpha)\% confidence interval around mean difference, standard error
}
\description{
Function to compute CI around the raw mean difference
}
\keyword{confidence}
\keyword{difference,}
\keyword{interval}
\keyword{mean}
