% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shieh.R
\name{shieh}
\alias{shieh}
\title{Function to compute CI around Shieh's effect size estimators}
\usage{
shieh(m1, m2, sd1, sd2, n1, n2, conf.level, unbiased, alternative)
}
\arguments{
\item{m1}{the average score of the first group}

\item{m2}{the average score of the second group}

\item{sd1}{the standard deviation the first group}

\item{sd2}{the standard deviation the second group}

\item{n1}{the first sample size}

\item{n2}{the second sample size}

\item{conf.level}{confidence level of the interval}

\item{unbiased}{a logical variable indicating whether to compute the biased or unbiased estimator.
If TRUE, unbiased estimator is computed (Hedges' g or Hedges' g'). Otherwise, bias estimator is computed (Cohen's d or Cohen's d').}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
Returns Cohen's estimators of effect size and (1-alpha)\% confidence interval around it, standard error
}
\description{
Function to compute CI around Shieh's effect size estimators
}
\keyword{Cohen's}
\keyword{confidence}
\keyword{effect}
\keyword{interval}
\keyword{sizes,}
